Modify input_processor program to dynamically resize memory
Realloc is now used to increase the memory for the char * pointer
that contains user input. This now allows for larger chunks of memory.
Over 300 characters have been tested in Valgrind with no leaks.
max_input variable was eliminated, as no limit is now required.
malloc for overall structure was also reduced to only allocate
enough space for 4 ints and one char pointer, as this is all that
is required, not the full size of the memory pointed to by the
char pointer, as previously thought.